+++ core/src/options/forwarded_field.rs
@@ -1,5 +1,5 @@
 use quote::{quote, ToTokens, TokenStreamExt};
-use syn::{Ident, Path};
+use syn::{Ident, Path, Type};
 
 use crate::{Error, FromField, FromMeta};
 
@@ -10,6 +10,8 @@ use super::ParseAttribute;
 pub struct ForwardedField {
     /// The ident of the field that will receive the forwarded value.
     pub ident: Ident,
+    /// The type of the field that will receive the forwarded value.
+    pub ty: Type,
     /// Path of the function that will be called to convert the forwarded value
     /// into the type expected by the field in `ident`.
     pub with: Option<Path>,
@@ -28,6 +30,7 @@ impl FromField for ForwardedField {
             ident: field.ident.clone().ok_or_else(|| {
                 Error::custom("forwarded field must be named field").with_span(field)
             })?,
+            ty: field.ty.clone(),
             with: None,
         };