+++ tests/compile-fail/generics_magic_field_missing_impl.rs
@@ -1,0 +1,8 @@
+#[derive(darling::FromDeriveInput)]
+#[darling(attributes(example))]
+pub struct Example {
+    // There is no FromGenerics impl for Vec<String>.
+    generics: Vec<String>,
+}
+
+fn main() {}